deployment/requirements.txt
streamlit
ultralytics
pillow
numpy
opencv-python-headless